// src/infrastructure/Enums/HealthPolicyType.ts
export enum HealthPolicyWeightType {
    WEIGHT = 'WEIGHT',
    HEIGHT = 'HEIGHT',
    BMI = 'BMI',
    BLOODTYPE = 'BLOODTYPE',
    GENDER = 'GENDER',
    AGE = 'AGE',
    POLICY_TYPE = 'POLICY_TYPE',
    CONSTANT = 'CONSTANT',
    BOOLEAN_TYPE = 'BOOLEAN_TYPE',

}

export default HealthPolicyWeightType;